+++ src/backend/core/tests/test_api_users.py
@@ -37,7 +37,7 @@ def test_api_users_list_authenticated():
     )
     assert response.status_code == 200
     content = response.json()
-    assert content["results"] == []
+    assert content == []
 
 
 def test_api_users_list_query_email():
@@ -58,22 +58,52 @@ def test_api_users_list_query_email():
         "/api/v1.0/users/?q=david.bowman@work.com",
     )
     assert response.status_code == 200
-    user_ids = [user["id"] for user in response.json()["results"]]
+    user_ids = [user["id"] for user in response.json()]
     assert user_ids == [str(dave.id)]
 
     response = client.get(
         "/api/v1.0/users/?q=davig.bovman@worm.com",
     )
     assert response.status_code == 200
-    user_ids = [user["id"] for user in response.json()["results"]]
+    user_ids = [user["id"] for user in response.json()]
     assert user_ids == [str(dave.id)]
 
     response = client.get(
         "/api/v1.0/users/?q=davig.bovman@worm.cop",
     )
     assert response.status_code == 200
-    user_ids = [user["id"] for user in response.json()["results"]]
+    user_ids = [user["id"] for user in response.json()]
     assert user_ids == []
+
+
+def test_api_users_list_limit(settings):
+    """
+    Authenticated users should be able to list users and the number of results
+    should be limited to 10.
+    """
+    user = factories.UserFactory()
+
+    client = APIClient()
+    client.force_login(user)
+
+    # Use a base name with a length equal 5 to test that the limit is applied
+    base_name = "alice"
+    for i in range(15):
+        factories.UserFactory(email=f"{base_name}.{i}@example.com")
+
+    response = client.get(
+        "/api/v1.0/users/?q=alice",
+    )
+    assert response.status_code == 200
+    assert len(response.json()) == 5
+
+    # if the limit is changed, all users should be returned
+    settings.API_USERS_LIST_LIMIT = 100
+    response = client.get(
+        "/api/v1.0/users/?q=alice",
+    )
+    assert response.status_code == 200
+    assert len(response.json()) == 15
 
 
 def test_api_users_list_query_email_matching():
@@ -94,13 +124,13 @@ def test_api_users_list_query_email_matching():
         "/api/v1.0/users/?q=alice.johnson@example.gouv.fr",
     )
     assert response.status_code == 200
-    user_ids = [user["id"] for user in response.json()["results"]]
+    user_ids = [user["id"] for user in response.json()]
     assert user_ids == [str(user1.id), str(user2.id), str(user3.id), str(user4.id)]
 
     response = client.get("/api/v1.0/users/?q=alicia.johnnson@example.gouv.fr")
 
     assert response.status_code == 200
-    user_ids = [user["id"] for user in response.json()["results"]]
+    user_ids = [user["id"] for user in response.json()]
     assert user_ids == [str(user4.id), str(user2.id), str(user1.id), str(user5.id)]
 
 
@@ -126,7 +156,7 @@ def test_api_users_list_query_email_exclude_doc_user():
     )
 
     assert response.status_code == 200
-    user_ids = [user["id"] for user in response.json()["results"]]
+    user_ids = [user["id"] for user in response.json()]
     assert user_ids == [str(nicole_fool.id)]
 
 
@@ -143,15 +173,15 @@ def test_api_users_list_query_short_queries():
 
     response = client.get("/api/v1.0/users/?q=jo")
     assert response.status_code == 200
-    assert response.json()["results"] == []
+    assert response.json() == []
 
     response = client.get("/api/v1.0/users/?q=john")
     assert response.status_code == 200
-    assert response.json()["results"] == []
+    assert response.json() == []
 
     response = client.get("/api/v1.0/users/?q=john.")
     assert response.status_code == 200
-    assert len(response.json()["results"]) == 2
+    assert len(response.json()) == 2
 
 
 def test_api_users_list_query_inactive():
@@ -166,7 +196,7 @@ def test_api_users_list_query_inactive():
     response = client.get("/api/v1.0/users/?q=john.")
 
     assert response.status_code == 200
-    user_ids = [user["id"] for user in response.json()["results"]]
+    user_ids = [user["id"] for user in response.json()]
     assert user_ids == [str(lennon.id)]